Refactor scanning functionality and UI improvements
Consolidate TWAIN and WIA backend handling into a unified structure within the `MainFrame` class. Introduce methods to get the current backend type and selected backend. Update the `Scan` method to utilize the selected backend, while simplifying the UI by removing separate tabs for TWAIN and WIA. Add new UI elements like a status strip and scan button, and improve image format handling with a new `ImageFormats` class. Enhance async/await usage for better responsiveness and update the `Dispose` method for proper resource cleanup. Overall, these changes improve maintainability and user experience.

Refactor scanner backend and update documentation

- Updated `Readme.md` with scanner setup instructions.
- Removed x64 configurations from `TownSuite.TwainScanner.sln`.
- Introduced new class structure in `IBackend.cs` for scanning and saving images in multiple formats with OCR support.
- Enhanced `Naps2TwainBackend.cs` and `OriginalTwainBackend.cs` for improved scanning functionality.
- Modified `OriginalWiaBackend.cs` for better WIA scanning management.
- Significant UI updates in `MainFrame.Designer.cs` to accommodate new features.
- Refactored `MainFrame.cs` to support dynamic backend selection (TWAIN, WIA, NAPS2).
- Made `Enabled` property public in `Ocr.cs` for better OCR management.
- Updated `Program.cs` to accept command-line arguments for backend selection.
- Changed `TwainScanner.csproj` to target .NET 8.0 and added necessary package references.

The Naps2TwainBackend is functional enough to scan images.

rename IBackend to ScannerBackends

Removed commented-out code

Enhance scanning and OCR processes with UI feedback

Updated `Naps2TwainBackend.cs` to include a progress bar and status label for user feedback during scanning. Restructured scanning logic with error handling using a `try-finally` block to ensure UI elements are reset after completion. Streamlined image processing within the scanning loop.

In `Ocr.cs`, added a status message for OCR processing and improved error handling with a `try-catch` block for user-friendly error messages. Finalized UI element management in a `finally` block to ensure proper reset after OCR completion.

Enhance Naps2Backend for driver flexibility and async ops

Updated Naps2Backend to support TWAIN and WIA drivers with a new constructor and conditional setup in ConfigureSettings. Refactored MainFrame methods to be asynchronous for improved responsiveness and error handling during scanning. Added localized error messages for better user guidance. Modified project file for .NET compatibility and cleaned up unused code in Scanner.cs. Overall, these changes improve functionality and user experience in the scanning application.

Refactor MainFrame layout and clean up ResX comments

Removed `this.` prefix from member variable declarations in
`MainFrame.Designer.cs` and adjusted control properties for
improved layout. Updated `ClientSize` and `MainMenu1`
dimensions. Simplified event handler assignments.

Formatted XML comments in `MainFrame.resx` for better
consistency and readability by removing unnecessary whitespace.

Update GitHub Actions workflow for .NET 8.0

Upgraded `actions/checkout` to v4 and replaced `microsoft/setup-msbuild` with `actions/setup-dotnet` for .NET version 8.0.x. Changed build command from `msbuild` to `dotnet build` with Release configuration. Updated artifact name from `win-x64` to `win-x86` and modified artifact paths to match new output structure.

cleanup unused code

Enhance scanning options and UI in TwainScanner

Updated ScanOptions to include BitDepth and Device properties.
Added ComboBox controls for DPI and color selection in the
scanning interface. Increased tpTWAINScan tab height to
accommodate new UI elements. Introduced DPIs and Colors
classes for managing options. Modified event handlers to
ensure correct device selection and backend changes on
tab switches. Minor formatting adjustments for consistency.
