+++ .github/workflows/dotnet.yml
@@ -10,18 +10,18 @@ jobs:
   windows-build:
     runs-on: windows-latest
     steps:
-    - uses: actions/checkout@v3
-    - name: Setup MSBuild.exe
-      uses: microsoft/setup-msbuild@v1.1
-    - name: Restore dependencies
-      run: dotnet restore TownSuite.TwainScanner.sln
+    - uses: actions/checkout@v4
+    - name: Setup .NET
+      uses: actions/setup-dotnet@v4
+      with:
+        dotnet-version: 8.0.x
     - name: Build
-      run: msbuild TownSuite.TwainScanner.sln -property:Configuration=Release
+      run: dotnet build TownSuite.TwainScanner.sln -c Release
 
     - name: Archive artifacts
       uses: actions/upload-artifact@v4
       with:
-        name: win-x64
+        name: win-x86
         path: |
-          TownSuite.TwainScanner/bin/Release
+          TownSuite.TwainScanner/bin/x86/Release
         retention-days: 1
